🚚 [Move]: #104 mock-data を test-fixtures へ移動 + 11 stories の import 置換
src/lib/mock-data/ は本来 lib/api/ の内部実装に閉じていたが、Storybook 11 件が
直接 import している。useProject 化で lib/api / mock-data を削除するため、
Storybook が依存する fixture を src/test-fixtures/ に分離して逃がす。

機械的に @/lib/mock-data → @/test-fixtures に書き換えのみ。fixture の中身
(initialColumns / initialTasks) は無変更。

Refs #104
