+++ src/features/detail/components/StatusSelect/index.stories.tsx
@@ -1,6 +1,6 @@
 // @jsdoc-rules-disable
 import type { Meta, StoryObj } from "@storybook/react-vite";
-import { initialColumns } from "@/lib/mock-data";
+import { initialColumns } from "@/test-fixtures";
 import { StatusSelect } from ".";
 
 const meta: Meta<typeof StatusSelect> = {
