+++ src/features/task-form/components/ParentTaskSelect/index.stories.tsx
@@ -1,6 +1,6 @@
 // @jsdoc-rules-disable
 import type { Meta, StoryObj } from "@storybook/react-vite";
-import { initialTasks } from "@/lib/mock-data";
+import { initialTasks } from "@/test-fixtures";
 import { ParentTaskSelect } from ".";
 
 const meta: Meta<typeof ParentTaskSelect> = {
