+++ src/features/task-form/components/TaskForm/TaskFormStatus/index.stories.tsx
@@ -1,6 +1,6 @@
 // @jsdoc-rules-disable
 import type { Meta, StoryObj } from "@storybook/react-vite";
-import { initialColumns } from "@/lib/mock-data";
+import { initialColumns } from "@/test-fixtures";
 import { TaskFormStatus } from ".";
 
 const meta: Meta<typeof TaskFormStatus> = {
